feat: replace amount on BigDecimal

+++ deposit-calculating-core/src/main/java/lekton/deniill/models/DepositModel.java
@@ -1,5 +1,6 @@
 package lekton.deniill.models;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -8,7 +9,7 @@ import lombok.Data;
 @AllArgsConstructor
 @Builder
 public class DepositModel {
-    private double amount;
-    private double percentage;
-    private double multiplier;
+    private BigDecimal amount;
+    private BigDecimal percentage;
+    private BigDecimal multiplier;
 }
